[DEVOPS-694] Move metrics to root
